+++ src/libs/satellite/src/db/impls.rs
@@ -1,4 +1,4 @@
-use crate::db::quota::init_rates;
+use crate::db::rates::init_rates;
 use crate::db::types::state::{DbHeapState, Doc, StableKey};
 use crate::SetDoc;
 use candid::Principal;
