+++ src/libs/satellite/src/db/mod.rs
@@ -1,7 +1,7 @@
 mod assert;
 pub mod impls;
 mod msg;
-mod quota;
+mod rates;
 mod state;
 pub mod store;
 pub mod types;
