+++ src/libs/satellite/src/db/rates.rs
